Migrate to Long and long timestamps

+++ lb-rest-task-scheduler/src/main/java/th/co/loxbit/rest_task_scheduler/scheduler/service/JobSchedulingService.java
@@ -8,12 +8,12 @@ public interface JobSchedulingService {
 
   int SERVICE_CODE = 7000;
 
-  void scheduleJob(int start, int end, String message, String owner);
+  void scheduleJob(long start, long end, String message, String owner);
 
   void descheduleJob(String jobId);
 
   List<GatewaySchedule> getScheduledJobs();
 
-  void updateJob(String jobId, int newStart, int newEnd, String newMessage, String newOwner);
+  void updateJob(String jobId, long newStart, long newEnd, String newMessage, String newOwner);
 
 }
